camera 임시구현
