+++ app/src/main/java/com/example/vibecapandroid/coms/HomeResponse.kt
@@ -1,2 +1,16 @@
 package com.example.vibecapandroid.coms
 
+data class CaptureResponse(
+
+    val is_success : Boolean,
+    val code : Int,
+    val message : String,
+    val CaptureResult : CaptureResult
+)
+
+data class CaptureResult(
+    val label : String,
+    val youtube_lnk : String,
+    val video_Id : String
+)
+
